pyproject.toml
[project]
name = "lse"
version = "0.2.0"
description = "The Live Service Engine handles execution and logging of search requests within Exquisitor."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "h5py>=3.13.0",
    "invoke>=2.2.0",
    "ipython>=9.2.0",
    "msgpack>=1.1.0",
    "ollama>=0.4.8",
    "open-clip-torch>=2.32.0",
    "orjson>=3.10.18",
    "pandas>=2.3.3",
    "pdoc>=16.0.0",
    "rangehttpserver>=1.4.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "scikit-learn>=1.6.1",
    "sphinx-autodoc-typehints>=3.2.0",
    "sphinx-rtd-theme>=3.0.2",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
    "zarr>=3.0.5",
]

[dependency-groups]
dev = [
    "licenseheaders>=0.8.8",
    "reuse>=5.1.1",
    "ruff>=0.11.12",
]
